// src/models/reconsillation-report.interface.ts
export interface ReconsillationReportInterface {
  afcTransactionType: string;
  afcPaymentMethod: string;
  reconStatus: string;
  transactionDate: string;
  transactionId: string;
  ppblTransactionType: string;
  ppblPaymentMethod: string;
  merchantName: string;
  orderId: string;
  issuingBank: string;
  paymentAmount: string;
  detailNo: string;
  merchantId: string;
  originalMerchantId: string;
  mbid: string;
  paymentType: string;
  paymentCurrency: string;
  responseCode: string;
  bankTxnId: string;
  userCustId: string;
  bankGateWay: string;
  status: string;
  bankFeeCurrency: string;
  bankFeeAmount: string;
  merchantCommission: string;
  serviceTax: string;
  convenienceFee: string;
  convenienceTax: string;
  settledDate: string;
  refundAmount: string;
  industry: string;
  settlementTxnId: string;
  instCode: string;
  externalSerialNo: string;
  reconId: string;
  settlementType: string;
  settleType: string;
  splitFlag: string;
  vouchercode: string;
  aggregatorMerchantId: string;
  prepaidCard: string;
  productCode: string;
  cardNetwork: string;
  instantSettlement: string;
  pwpCategory: string;
  feeFactor: string;
  acquiringFee: string;
  platformFee: string;
  acquiringTax: string;
  platformTax: string;
  addMoneyType: string;
  corporateCard: string;
  utr: string;
  firstSixDigits: string;
  lastFourDigits: string;
  payerMaskedCardNo: string;
  maskedTokenNo: string;
  dccId: string;
  merchantSolutionType: string;
  splitFieldDlLastUpdated: string;
  bankTid: string;
  apprCode: string;
  subWalletFeeDetails: string;
  subWalletTaxDetails: string;
  conveniencePlatformFee: string;
  conveniencePlatformFeeServiceTax: string;
  orderSuccessTime: string;
  stan: string;
  runType: string;
  runDate: string;
  splitType: string;
  splitDelay: string;
  upiModeSubType: string;
  dlLastUpdated: string;
  insertTime: string;
}